# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Planner API documentation",
        "description": "API Documentation for Travel Planner",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/destination": {
            "get": {
                "tags": [
                    "Destination"
                ],
                "summary": "Get all destinations for logged in user",
                "description": "Get all destinations for logged in user",
                "operationId": "get_destination",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Destination"
                ],
                "summary": "Add a destination for a trip",
                "description": "Add a destination to trip after creation",
                "operationId": "create_destination",
                "parameters": [
                    {
                        "name": "trip_id",
                        "in": "query",
                        "description": "required | exists:trips,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "required | string | max:100",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "date_format:Y-m-d | after:today",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "date_format:Y-m-d | after:start_date",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "cost",
                        "in": "query",
                        "description": "numeric | min:0",
                        "schema": {
                            "type": "decimal"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/destination/update": {
            "post": {
                "tags": [
                    "Destination"
                ],
                "summary": "Update destination",
                "description": "Update destination",
                "operationId": "update_destination",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "required | exists:destinations,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "string | max:100",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "date_format:Y-m-d | after:today",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "date_format:Y-m-d | after:start_date",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "cost",
                        "in": "query",
                        "description": "numeric | min:0",
                        "schema": {
                            "type": "decimal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/destination/delete": {
            "post": {
                "tags": [
                    "Destination"
                ],
                "summary": "Delete destination",
                "description": "Delete destination",
                "operationId": "delete_destination",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "required | exists:destinations,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}